Refactored to break dependency cycle.
